minro class rename issue
